requirements.txt
fastapi
uvicorn
cadquery
groq
python-dotenv
python-multipart
trimesh
numpy